Added blank DB
added a blank DERBY database called ScedulingAppDB. This will be our main database. Database should be moved to your derby folder on your computer. When finished updating database or editing it the new version must be copied back into the project folder for github so that everyone get the updated version